refactor: introduce separate files for shells
